requirements.txt
crewai
crewai_tools
langchain-groq
python-dotenv
yt_dlp
youtube-transcript-api